Changes in design choices and integral constants.
Mention the changes in the design choices due to the recent WIP support of C++14 `constexpr` for mutable operations, and remove the old paragraph that mention the explicit restrictions that have been lifted.

Introduce `constant`. It's not the most useful part of the library but it lacked a bit of exposure since it is quite different from the rest of the library.